fix(tests): on python 3.8
